questionaire 1.0
